requirements.txt
fastapi
uvicorn
pillow
torch
opencv-python
python-multipart
starlette
pandas
requests
IPython
torchvision
pyyaml
tqdm
matplotlib
seaborn

